add arcade token template
