+++ packages/sdk/src/templates/index.ts
@@ -1,1 +1,2 @@
 export * from './stablecoin';
+export * from './arcadeToken';
